[Core] Changes in applicators architecture
